requirements.txt
# Telegram Bot
aiogram==3.17.0
aiohttp==3.10.10

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
pgvector==0.3.6

# OpenAI
openai==1.57.0

# Web API
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-dateutil==2.9.0
pytz==2024.2

# PDF/Document processing
pypdf==5.1.0
python-docx==1.1.2
Pillow==11.0.0

# URL parsing
beautifulsoup4==4.12.3
httpx==0.28.1
yt-dlp==2024.12.6

# Voice processing (Whisper)
# Note: For production, consider using OpenAI's Whisper API instead
# openai-whisper==20240930

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Environment
python-dotenv==1.0.1
